added security based on roles

+++ src/main/java/com/restproj/controller/NoteController.java
@@ -1,18 +1,27 @@
 package com.restproj.controller;
 
 import com.restproj.model.Note;
+import com.restproj.model.Role;
+import com.restproj.model.User;
 import com.restproj.service.NoteService;
+import com.restproj.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @RestController
-@RequestMapping("/notes")
+@RequestMapping("/api/notes")
 public class NoteController {
     @Autowired
     private NoteService noteService;
+
+    @Autowired
+    private UserService userService;
 
     @GetMapping
     public List<Note> getAll(){
@@ -20,29 +29,53 @@ public class NoteController {
     }
 
     @GetMapping("/{id}")
-    public Note getById(@PathVariable(name = "id") Long id){
-        return noteService.findById(id);
+    public ResponseEntity<Note> getById(@PathVariable(name = "id") Long id, Principal principal){
+        User user = userService.findByLogin(principal.getName());
+        Note note = noteService.findById(id);
+        if(note == null){
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
+        if(user.equals(note.getUser()) || user.getRole().equals(Role.ADMIN)) {
+            return new ResponseEntity<>(noteService.findById(id), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Note create(@RequestBody Note note){
+    public Note create(@RequestBody Note note, Principal principal){
+        User user = userService.findByLogin(principal.getName());
+        note.setUser(user);
         return noteService.save(note);
     }
 
     @PutMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public Note update(@PathVariable Long id, @RequestBody Note note){
+    public ResponseEntity<Note> update(@PathVariable Long id, @RequestBody Note note, Principal principal){
         if(!id.equals(note.getId())){
-            throw new IllegalStateException("Path variable id and real note id are different!");
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
-        return noteService.update(note);
+        User user = userService.findByLogin(principal.getName());
+        Note oldNote = noteService.findById(id);
+        if(oldNote == null){
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
+        if (user.equals(oldNote.getUser()) || user.getRole().equals(Role.ADMIN)){
+            return new ResponseEntity(noteService.update(note), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Long id){
+    public ResponseEntity delete(@PathVariable Long id, Principal principal){
+        User user = userService.findByLogin(principal.getName());
         Note note = noteService.findById(id);
-        noteService.delete(note);
+        if(note == null){
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
+        if(user.equals(note.getUser()) || user.getRole().equals(Role.ADMIN)){
+            noteService.delete(note);
+            return new ResponseEntity(HttpStatus.OK);
+        }
+        return new ResponseEntity(HttpStatus.FORBIDDEN);
     }
 }
